requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
werkzeug==2.3.7
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.0.0
gunicorn==21.2.0

